added return status to actions and reaper

+++ utils.c
@@ -58,7 +58,7 @@ void	ph_putstr_fd(const char *s, int fd)
 			return ;
 }
 
-void	*ph_calloc(size_t count, size_t size)
+void	*ph_calloc(size_t count, uint_fast32_t size)
 {
 	void			*mem;
 	unsigned char	*ptr;
